Moved many more widgets Merged widgets_integrated and widget_standalone into one widgets folder Moved all kivy_soil modules into new kivy_soil package
