+++ mmplayer/kivy_soil/app_recycleview/viewclass.py
