+++ mmplayer/layouts/pc_layout.kv
@@ -1,15 +1,15 @@
 #: include layouts/app_style.kv
-#: import MultiLineLabel widgets_standalone.multi_line_label.MultiLineLabel
+#: import MultiLineLabel widgets.multi_line_label.MultiLineLabel
 #: import SideBarRecycleView rv_sidebar.sidebar.SideBarRecycleView
 #: import PlayBackBar playback_bar.playback_bar.PlayBackBar
-#: import ResizableBoxLayout widgets_integrated.resizable_boxlayout.ResizableBoxLayout
-#: import SmallVideo widgets_integrated.small_video.SmallVideo
-#: import InfoWidget widgets_standalone.info_widget.InfoWidget
+#: import ResizableBoxLayout widgets.resizable_boxlayout.ResizableBoxLayout
+#: import SmallVideo widgets.small_video.SmallVideo
+#: import InfoWidget kivy_soil.info_widget.InfoWidget
 #: import AppMenu other.garden_contextmenu.AppMenu
 #: import AppMenuTextItem other.garden_contextmenu.AppMenuTextItem
 #: import ContextMenu other.garden_contextmenu.ContextMenu
 #: import ContextMenuTextItem other.garden_contextmenu.ContextMenuTextItem
-#: import TerminalWidget widgets_standalone.terminal_widget.TerminalWidget
+#: import TerminalWidget widgets.terminal_widget.TerminalWidget
 
 
 <Jotube>:
