+++ mmplayer/media_controller/controller.kv
@@ -1,4 +1,4 @@
-#: import LabelButton widgets_integrated.labelbutton.LabelButton
+#: import LabelButton widgets.labelbutton.LabelButton
 
 <MediaButton>:
     bg_color: 0.20, 0.20, 0.40, 1
