+++ mmplayer/widgets/background_label.py
