+++ mmplayer/widgets/background_stacklayout.py
