+++ mmplayer/widgets/popup2.py
@@ -1,10 +1,10 @@
+from kivy_soil.hover_behavior import HoverBehavior
+from kivy.uix.modalview import ModalView
 from kivy.uix.popup import Popup
 from kivy.lang import Builder
-from kivy.uix.modalview import ModalView
-from behaviors import hover_behavior
 
 
-class AppPopup(hover_behavior.HoverBehavior, Popup):
+class AppPopup(HoverBehavior, Popup):
     hover_height = 20
 
     def open(self):
