+++ mmplayer/widgets/small_video.py
